requirements.txt
GitPython==3.1.44
requests==2.32.4
python-dotenv==1.1.0
matplotlib==3.10.3
Pillow==11.2.1
pathspec==0.12.1